Add a test about saving POJO only annotated with @Id
This test has been adapted from https://github.com/jeanlaurent/JongoId
Jongo allows a POJO with a String field only annotated with @Id to be
inserted even if it is empty.
According to the documentation it should not be possible : the field
must be annotated with @ObjectId
Nevertheless this is easy to misuse so we have to support it.

refs #201
refs #199